fix bug hieen thi null

+++ DMS/Rpc/kpi-general/KpiGeneralController.cs
@@ -186,7 +186,7 @@ namespace DMS.Rpc.kpi_general
             {
                 KpiCriteriaGeneralId = x.Id,
                 KpiCriteriaGeneral = new KpiGeneral_KpiCriteriaGeneralDTO(x),
-                KpiGeneralContentKpiPeriodMappings = KpiPeriods.ToDictionary(x => x.Id, y => (long?)null),
+                KpiGeneralContentKpiPeriodMappings = KpiPeriods.ToDictionary(x => x.Id, y => (decimal?)null),
                 KpiGeneralContentKpiPeriodMappingEnables = KpiPeriods.ToDictionary(x => x.Id, y => ((y.Id < KpiGeneral_KpiGeneralDTO.CurrentMonth.Id && y.Id <= Enums.KpiPeriodEnum.PERIOD_MONTH12.Id) 
                                                                                                 || (y.Id > Enums.KpiPeriodEnum.PERIOD_MONTH12.Id && y.Id < KpiGeneral_KpiGeneralDTO.CurrentQuarter.Id && y.Id <= Enums.KpiPeriodEnum.PERIOD_QUATER04.Id)) == true
                                                                                                 ? false : true),
